0.0.0.0 - pre_render

+++ core/mesh.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "node.hpp"
+
 class Mesh : public Node3D {
 public:
 	unsigned int VAO, VBO, EBO;
@@ -7,7 +9,7 @@ public:
 
 	Mesh(
 		std::vector<float> vertices,
-		std::vector<float> indices,
+		std::vector<unsigned int> indices,
 		std::string new_name = "Mesh",
 		glm::vec3 new_position = glm::vec3(0.0f, 0.0f, 0.0f),
 		glm::vec3 new_rotation = glm::vec3(0.0f, 0.0f, 0.0f),
